MAJOR REPO CHANGE will probably break OLD.
Changes Made:
MAIN\ consist of the new project files, MAIN\Project\ is the main directory to the DEV4_Project itself.
OLD\ is the prior version of the project before making it compatible with CMake to build on any PC.

In the project itself, comments were modified and a few fixes/clean ups were done. After cloning the old project to my home PC, I was unable to run the project which lead me to believe that I was doing the project wrong from the start. [When I started on this project, I built it using Cmake and then modified the project within that .SLN file.]